+++ game_patch/rf/particle_emitter.h
@@ -137,7 +137,7 @@ struct ParticleEmitter
     float time_to_change;
     float current_state_time;
     bool active;
-    int field_144;
+    float active_distance;
     ParticleEmitter *next;
     ParticleEmitter *prev;
     ParticleEmitter *next_entity_emitter;